requirements.txt
cryptography==2.3
pyjwt==1.5.3
paho-mqtt==1.3.1
tendo==0.2.12
